Create FitnessData.Rmd
This project was apart of the Google Data Analytics Capstone project. The main goal of this project was to draw insights and trends from Fitbit users (smart watch company) so we can improve the products of BellaBeat Company. Analysis process: Ask, Prepare, Process, Analyze, Share, and Act. This project was done in R using the Tidyverse library.